+++ lib/instructions/instruction_metrics.c
@@ -1322,7 +1322,7 @@ const struct evmc_instruction_metrics* evmc_get_instruction_metrics_table(
     switch (revision)
     {
     case EVMC_ISTANBUL:
-        /* Currently the same as Costantinople */
+    case EVMC_CONSTANTINOPLE2:
     case EVMC_CONSTANTINOPLE:
         return constantinople_metrics;
     case EVMC_BYZANTIUM: