+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   The **Istanbul** EVM revision.
 - Added: [[#182](https://github.com/ethereum/evmc/pull/182)]
   The `is_zero()` C++ helper for basic data types.
+- Added: [[#186](https://github.com/ethereum/evmc/pull/186)]
+  Reserved the post-Constantinople EVM revision number.
 - Deprecated: [[#184](https://github.com/ethereum/evmc/pull/184)]
   The `EVMC_LATEST_REVISION` is deprecated, replaced with `EVMC_MAX_REVISION`.